requirements.txt
discord.py-self
pydantic
python-dotenv
sentry-sdk